Added buttons use default "chatbot@example.com" e-mail for chatbot user account
