+++ chat/hooks.py
@@ -75,6 +75,8 @@ web_include_js = ['chat.bundle.js'] if is_frappe_above_v13 else [
 # before_install = "chat.install.before_install"
 after_install = "chat.patches.migrate_chat_data.execute"
 
+before_uninstall = "chat.patches.migrate_chat_data.delete_default_chatbot_user"
+
 # Desk Notifications
 # ------------------
 # See frappe.core.notifications.get_notification_config
