+++ chat/public/js/components/chat_list.js
@@ -138,12 +138,12 @@ export default class ChatList {
     });
 
     $('.add-ai-room').on('click', function (e){
-      frappe.call('chat.api.room.direct_room_exists',{users:[frappe.session.user,'ai@help.ai']})
+      frappe.call('chat.api.room.direct_room_exists',{users:[frappe.session.user,'chatbot@example.com']})
         .then(response => {
           if(!response.message){// If direct room doesn't already exist, create one
-            create_private_room('AI Helper',['ai@help.ai'],'Direct')
+            create_private_room('Chatbot',['chatbot@example.com'],'Direct')
           }
-          $('.avatar[title|=\'ai@help.ai\'').trigger('click') // Click to navigate to AI chat room
+          $('.avatar[title|=\'chatbot@example.com\'').trigger('click') // Click to navigate to AI chat room
         })
         .catch(ex => {
           console.log(ex)
